Belgian motor third party liability insurance portfolio

Dataset containing the number of motor insurance claims and several characteristics for a portfolio of around 163000 policyholders.

mtpl_be

Format

A data frame with 163210 rows and 13 variables:

id

policyholder id

nclaims

number of claims

expo

exposure period, as a fraction of a year

coverage

coverage type: TPL, TPL+ or TPL++

sex

gender: female or male

fuel

fuel type: diesel or gasoline

use

use type: private or work

fleet

fleet vehicle: 0 or 1

ageph

age of the policyholder, in years

bm

bonus-malus level, higher is worse

power

horsepower of the vehicle, in kilowatt

agec

age of the vehicle, in years

postcode

first two digits of the postal code of the municipality of residence